// src/interfaces/IPlanets.ts
export interface IPlanets {
    count:    number;
    next:     string;
    previous: null;
    results:  IPlanetsResult[];
}

export interface IPlanetsResult {
    name:            string;
    rotation_period: string;
    orbital_period:  string;
    diameter:        string;
    climate:         string;
    gravity:         string;
    terrain:         string;
    surface_water:   string;
    population:      string;
    residents:       string[];
    films:           string[];
    created:         Date;
    edited:          Date;
    url:             string;
}